use event emitter to rewrite terminate ws
in the past we use another field in M to communicate. and the video ws side periodically check that variable.
this worked because video ws periodically run a function to proxy frames, we just let it do some extra work.

however this doesn't make sense for audio & ctrl ws -- we don't know when there's data in.
so it's time to change our design -- use event emitter, the classical Observer Pattern.